+++ source/matrix/minor_matrix.c
@@ -12,20 +12,20 @@
 
 #include <header.h>
 
-float	minor_matrix3(t_matrix3 src, size_t x_del, size_t y_del)
+double	minor_matrix3(t_matrix3 src, size_t x_del, size_t y_del)
 {
 	t_matrix2	mtx;
-	float		minor;
+	double		minor;
 
 	mtx = sub_matrix3(src, x_del, y_del);
 	minor = determinant_matrix2(mtx);
 	return (minor);
 }
 
-float	minor_matrix4(t_matrix4 src, size_t x_del, size_t y_del)
+double	minor_matrix4(t_matrix4 src, size_t x_del, size_t y_del)
 {
 	t_matrix3	mtx3;
-	float		minor;
+	double		minor;
 
 	mtx3 = sub_matrix4(src, x_del, y_del);
 	minor = determinant_matrix3(mtx3);
