+++ source/tuple/scalar_product_vector.c
@@ -17,7 +17,7 @@
 /*   Ou "dot product".														  */
 /*                                                                            */
 /* ************************************************************************** */
-float	scalar_product_vector(t_tuple *a, t_tuple *b)
+double	scalar_product_vector(t_tuple *a, t_tuple *b)
 {
 	return ((a->x * b->x) + (a->y * b->y) + (a->z * b->z) + (a->w * b->w));
 }
